[FE-127] Moved the Redeem componenet to the Buttons component.
